moved around all the home-page code and included working home-page test case
