+++ unest-app/src/App.js
@@ -3,7 +3,9 @@ import "./App.css";
 import {Route, Routes} from "react-router-dom";
 import {BrowserRouter} from "react-router-dom";
 import SignUpPage from "./components/SignUpPage";
+import Properties from "./components/Properties";
 import RegistrationInput from "./components/RegistrationInput";
+
 
 
 function App(){
@@ -11,6 +13,7 @@ function App(){
         <BrowserRouter>
             <Routes>
                 <Route index element={<SignUpPage />} />
+                <Route path="/homepage" element={<Properties/>} />
             </Routes>
         </BrowserRouter>
     )
